Refactor code structure (#13)
* Modify the file with settings for pre-commit hooks

* Fix requirement files

* Modify settings in configuration files

* Fix .gitignore file

* Fix annotations + small style fixes

* Delete output folders and small style fix in train file

* Move tests to separate folder

* Small style fixes

* Write the script to download and organize FER+ data

* Refactor code structure

* Fix paths

* Add test images

* Add ability to download video for conducting corresponding tests
