+++ .pre-commit-config.yaml
@@ -1,47 +1,52 @@
-exclude: _pb2\.py$
 repos:
-- repo: https://github.com/psf/black
-  rev: 20.8b1
-  hooks:
-    - id: black
--   repo: https://github.com/timothycrosley/isort
-    rev: 5.5.3  # pick the isort version you'd like to use from https://github.com/timothycrosley/isort/releases
+  - repo: https://github.com/asottile/pyupgrade
+    rev: v2.19.4
     hooks:
-    - id: isort
-- repo: https://github.com/asottile/yesqa
-  rev: v1.1.0
-  hooks:
-  - id: yesqa
-    additional_dependencies:
-    - flake8-bugbear==20.1.4
-    - flake8-builtins==1.5.2
-    - flake8-comprehensions==3.2.2
-    - flake8-tidy-imports==4.1.0
-    - flake8==3.7.9
--   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v2.3.0
+      - id: pyupgrade
+        args: [ "--py36-plus" ]
+  - repo: https://github.com/pre-commit/mirrors-isort
+    rev: 1ba6bfc # Use the revision sha / tag you want to point at
     hooks:
-    - id: check-docstring-first
-    - id: check-json
-    - id: check-merge-conflict
-    - id: check-yaml
-    - id: debug-statements
-    - id: end-of-file-fixer
-    - id: trailing-whitespace
-    - id: flake8
-    - id: requirements-txt-fixer
-- repo: https://github.com/asottile/pyupgrade
-  rev: v1.13.0
-  hooks:
-  - id: pyupgrade
-    args: ['--py36-plus']
-- repo: https://github.com/pre-commit/pygrep-hooks
-  rev: v1.5.1
-  hooks:
-  - id: python-check-mock-methods
-  - id: python-use-type-annotations
-- repo: https://github.com/pre-commit/mirrors-mypy
-  rev: 9feadeb
-  hooks:
-  - id: mypy
-    args: [--ignore-missing-imports, --warn-no-return, --warn-redundant-casts, --disallow-incomplete-defs]
+      - id: isort
+        args: ["--profile", "black"]
+  - repo: https://github.com/psf/black
+    rev: 21.12b0
+    hooks:
+      - id: black
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: 4.0.1
+    hooks:
+      - id: flake8
+        language_version: python3
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.0.1
+    hooks:
+      - id: check-docstring-first
+      - id: check-json
+      - id: check-merge-conflict
+      - id: check-yaml
+      - id: debug-statements
+      - id: end-of-file-fixer
+      - id: trailing-whitespace
+      - id: requirements-txt-fixer
+  - repo: https://github.com/pre-commit/pygrep-hooks
+    rev: v1.9.0
+    hooks:
+      - id: python-check-mock-methods
+      - id: python-use-type-annotations
+      - id: python-check-blanket-noqa
+      - id: python-use-type-annotations
+      - id: text-unicode-replacement-char
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: 9feadeb
+    hooks:
+      - id: mypy
+        exclude: ^tests/
+        args:
+          [
+              --disallow-untyped-defs,
+              --check-untyped-defs,
+              --warn-redundant-casts,
+              --no-implicit-optional,
+              --strict-optional
+          ]
