+++ fer_pytorch/get_data.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+mkdir dataset  && \
+cd dataset  && \
+wget https://github.com/Emilien-mipt/FERplus-Pytorch/releases/download/0.0.2/fer_data.zip && \
+unzip fer_data.zip  && \
+rm fer_data.zip
