+++ pyproject.toml
@@ -1,10 +1,3 @@
-# Example configuration for Black.
-
-# NOTE: you have to use single-quoted strings in TOML for regular expressions.
-# It's the equivalent of r-strings in Python.  Multiline strings are treated as
-# verbose regular expressions by Black.  Use [ ] to denote a significant space
-# character.
-
 [tool.black]
 line-length = 119
 target-version = ['py36', 'py37', 'py38']
@@ -16,7 +9,7 @@ exclude = '''
   | \.hg
   | \.mypy_cache
   | \.tox
-  | \.cassava_venv
+  | \.venv
   | _build
   | buck-out
   | build
