--- requirements/requirements-dev.txt
@@ -1,13 +1,0 @@
--r requirements.txt
-
-tensorboard>=2.5.0,<=2.8.0
-torch_lr_finder==0.2.1
-
-# repo maintenance tooling
-black>=20.8b1,<=22.1.0
-flake8>=3.9.0,<=4.0.1
-isort>=5.8.0.<=5.10.1
-mypy>=0.812,<=0.941
-
-# testing requirements
-pytest==7.0.1
