+++ setup.cfg
@@ -1,10 +1,13 @@
 [flake8]
-max-line-length = 119
-exclude =.git,__pycache__,docs/source/conf.py,build,dist
-ignore = I101,I201,F401,F403,S001,D100,D101,D102,D103,D104,D105,D106,D107,D200,D205,D400,W504,E203,D202,D401,W503
+max-line-length = 120
+exclude =.git,__pycache__,docs/source/conf.py,build,dist,tests
+ignore = I101,D100,D101,D102,D103,D104,D105,D107,D401,E203,I900,N802,N806,N812,W503,S311,S605,S607
 
 [mypy]
 ignore_missing_imports = True
-
-[isort]
-profile = black
+disallow_untyped_defs = True
+check_untyped_defs = True
+warn_redundant_casts = True
+no_implicit_optional = True
+strict_optional = True
+exclude = tests
